requirements.txt
streamlit==1.19.0
pandas
scikit-learn
scikit-surprise